fix(game-tests): 收敛 Game.Tests 测试告警
- 修复多组 YAML 与 persistence 测试中的 ConfigureAwait 使用与状态校验，清理低风险 analyzer 告警

- 重构 PersistenceTestUtilities 为单类型文件，消除测试辅助模型的文件命名告警

- 更新 analyzer-warning-reduction 跟踪与 trace，记录 RP-054 批次结果与 75 文件阈值停止点
